allow decimal values for `amount`
When users supply `amount` for features X PU in a legacy project to be
imported, this column can hold arbitrary numeric values (such as areas,
etc.) other than integer counts, so decimal values should be allowed.

+++ api/apps/geoprocessing/src/migrations/geoprocessing/1653650018540-AddAmountFromLegacyProjectToScenarioFeaturesData.ts
@@ -5,7 +5,7 @@ export class AddAmountFromLegacyProjectToScenarioFeaturesData1653916456540
   public async up(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(`
       ALTER TABLE scenario_features_data
-        ADD COLUMN amount_from_legacy_project int;
+        ADD COLUMN amount_from_legacy_project double precision;
     `);
   }
